% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_intervals.R
\name{index_difference_ci}
\alias{index_difference_ci}
\title{Confidence Interval for the Difference of Two Indices}
\usage{
index_difference_ci(
  env1,
  env2,
  seg_index = c("D", "Gini", "Mutual_info", "Atkinson", "Theil"),
  method = c("Asymp", "CBoot", "TBoot", "CBayBoot"),
  framework = "fm",
  n_b = 5000,
  confidence_level = 0.95,
  prior1 = NULL,
  prior2 = NULL
)
}
\arguments{
\item{env1}{nxk1 \code{matrix} of \code{numeric}.}

\item{env2}{nxk1 \code{matrix} of \code{numeric}.}

\item{seg_index}{\code{string} or \code{vector}. Which segregation indices
should be analyzed? The function works for any index available in SISeg
and any index the user manually added.}

\item{method}{\code{string} or \code{vector} indicating which method(s)
should be used. The methods can be \code{Asymp}, \code{Boot}, \code{CBoot}, \code{TBoot},
\code{BayModel}, \code{CBayBoot}.}

\item{framework}{\code{string}. One of \code{fm}, \code{ig}, \code{iu}.}

\item{n_b}{\code{numeric}. How many simulations should be returned? Used by
the bootstrap and Bayesian methods.}

\item{confidence_level}{\code{numeric} between 0 and 1 setting the confidence
level for the intervals.}

\item{prior1}{\code{list} of priors to be passed to the Bayesian models for
of the \code{env1}. See \code{\link{r_ig}}, \code{\link{r_iu}} and
\code{\link{r_fm}}.}

\item{prior2}{\code{list} of priors to be passed to the Bayesian models for
of the \code{env2}. See \code{\link{r_ig}}, \code{\link{r_iu}} and
\code{\link{r_fm}}.}
}
\value{
\code{list} of \code{lists} containing the confidence intervals and
point estimates. The \code{list} is ordered by method and then by
index.
}
\description{
The \code{index_difference_ci} function estimates different kinds of
confidence intervals for the difference between indices
in two different datasets, under a given framework. The function returns the
confidence intervals and point estimates in a list. It can calculate the
confidence intervals for more than one segregation index at the time. It
assumes that the two data creation processes were independent.
}
\details{
\code{index_difference_ci} starts from data, in the form of two
\code{matrices}, and calculates the confidence interval and point estimates
for difference of segregation indices in the two datasets. There are two main
parameters to be chosen. It assumes that the two data creation processes
were independent.

First, all estimations are always done under a framework, representing the
data generating process. The available frameworks are
\itemize{
\item{\code{Independent_groups} ('ig')}
\item{\code{Full_Multinomial} ('fm') -- corresponding to standard non-parametric
bootstrap}
\item{\code{Independent_Units} ('iu')}
}
The user can use the full names or the acronym. Framework names are
case-sensitive. By default the function will pick the 'Full_Multinomial'
framework, corresponding to standard bootstrap. The other frameworks
represent stratified sampling by group ('ig') or by units ('iu') with
proportionate allocations. If none of the frameworks represent the actual
data generating process 'fm' is the safest choice and it is usually
conservative with respect to 'ig' and 'iu'.

The second choice the user needs to make is about the \code{method} argument.
\code{index_difference_ci} can use six different methods to construct
confidence intervals and point-estimates:
\describe{
\item{\code{Asymp}}{Asymptotic confidence interval based on the delta method.}
\item{\code{Boot}}{Bootstrap percentile confidence interval, with no bias
correction.}
\item{\code{CBoot}}{Bootstrap percentile confidence interval, with bias
correction.}
\item{\code{TBoot}}{Studentized Bootstrap percentile confidence interval.}
\item{\code{BayModel}}{Credible interval and point-estimate based on two simple
conjugate Bayesian models.}
\item{\code{CBayBoot}}{Credible interval and point-estimate based on two simple
conjugate Bayesian models of the index bias (as opposed to a model of the
whole environment) in the two environments.}
}
Each confidence interval is linked to a specif point-estimate. The methods
\code{Asymp}, \code{Boot}, \code{TBoot} are based on the plug-in estimate of the segregation
index on the data. \code{CBoot} applies a
bootstrap correction to the plug-in point-estimate in order to diminish the
bias. \code{BayModel} draws upon two estimated Bayesian model (one per dataset)
and use as point-estimate the expected value of the indices' difference based
on the models.
\code{CBayBoot} creates two Bayesian models (one per environment) of the index
plug-in's bias; it creates the point-estimate of the difference between the
the bias-corrected plug-in estimates and use the models' percentile to create
credible intervals.

All the methods and frameworks will be substantially equivalent in large
samples, but in small and mid-size samples they can give very different
answers. Based on what we know about the estimation of indices, the
reccomandation is to use \code{CBayBoot} in small samples.
In all other cases, \code{CBoot} provides the best point estimate.
Finally, in large and mid-size samples, \code{TBoot} seems the confidence
interval with coverage closest to nominal. In all cases, \code{Asymp} and \code{Boot}
should be avoided.
}
\section{Warnings}{


In small samples, it is common to receive warnings about few failure in the
bootstrap or Bayesian estimations. The warnings are due to some simulated
environments having a row (or a column) with only zeros. In these cases, the
methods can fail in calculating the standard deviation or even the index
value. These cases will therefore be discarded in the inference procedure and
the function will produce warnings about this. However, this will be hardly
consequential for the final results.
}

\examples{
env <- matrix(seq(8), nrow = 2, byrow = TRUE)
env2 <- matrix(seq(9,16), nrow = 2, byrow = TRUE)
# confidence interval for a whole lot of indices and methods, just 1 line
ci_lot <- index_difference_ci(env, env2, n_b = 5000, seg_index = c("Theil",
    "D", "Gini", "Atkinson", "Mutual_info"), method =  c("Asymp", "CBoot",
    "CBayBoot", "TBoot"))
# D index, CBoot (notice the lower case)
ci_lot[["CBoot"]][["d"]]
# Atkinso Asymptotic
ci_lot[["Asymp"]][["atkinson"]]
# Produce warnings
set.seed(1233)
ci1 <- index_difference_ci(env, env2, n_b = 250, seg_index =
    c("Mutual_info"), method = c("CBoot", "CBayBoot", "TBoot"))
# Doesn't produce warnings
set.seed(1230)
ci2 <- index_difference_ci(env, env2, n_b = 250, seg_index =
    c("Mutual_info"), method = c("CBoot", "CBayBoot", "TBoot"))
# Very similar estimates
ci1$CBoot$mutual_info - ci2$CBoot$mutual_info
}
\references{
Allen, Rebecca, Simon Burgess, Russell Davidson, and Frank Windmeijer. 2015.
“More Reliable Inference for the Dissimilarity Index of Segregation.”
The Econometrics Journal 18 (1): 40–66.
}
\seealso{
Other ci: 
\code{\link{index_ci}()}
}
\concept{ci}
