% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul.R
\name{r_full_multinomial}
\alias{r_full_multinomial}
\title{Simulate Full Multinomial Framework}
\usage{
r_full_multinomial(n, p = F, dir_alpha = F)
}
\arguments{
\item{n}{\code{numeric} representing the number of individuals in the sample.}

\item{p}{\code{vector} of probability for the multinomial distribution.}

\item{dir_alpha}{\code{vector} the shape parameter of the Dirichlet-multinomial
distribution.}
}
\value{
\code{matrix} containing the generated sample
}
\description{
\code{r_full_multinomial} samples from one
distribution or, following a simple conjugate model, a Dirichlet-
multinomial. Returns the sample in the form of a matrix. At least, one
among p and dir must be specified.
}
\details{
Simulate one sample from from a multinomial distribution or
Dirichlet-multinomial distribution. The multinomial distribution will take as
probability vectors p. The Dirichlet-multinomial distribution will
take as alpha vectors dir_alpha. In case p is specified,
dir_alpha will be ignored and the sample will be simulated from a multinomial
distribution. p does not have to be normalized.

At least, one among p and dir_alpha must be specified.
}
\examples{
env_ <- matrix(c(1,2,3,4,5,6,7,8), nrow = 2, byrow = T)
n <- sum(env)
p <- c(t(env))
samp <- r_full_multinomial(n = n, p = p)
samp2 <- r_full_multinomial(n = n, dir_alpha = p)
all.equal(sum(env), sum(samp), sum(samp2))
}
\seealso{
Other frameworks: 
\code{\link{r_independent_group}()},
\code{\link{r_independent_unit}()}
}
\concept{frameworks}
