% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{seg_ind}
\alias{seg_ind}
\alias{d_ind}
\alias{gini_ind}
\alias{mutual_info_ind}
\alias{theil_ind}
\alias{atkinson_ind}
\alias{v_ind}
\alias{isolation_ind}
\alias{isolationInv_ind}
\title{Segregation Indices}
\usage{
d_ind(env)

gini_ind(env)

mutual_info_ind(env)

theil_ind(env)

atkinson_ind(env)

v_ind(env)

isolation_ind(env)

isolationInv_ind(env)
}
\arguments{
\item{env}{2xN matrix of numeric.}
}
\value{
\code{numeric} representing the index value for the data.
}
\description{
All these functions calculate an index for the input environment, which is
represented as a matrix. They output a positive \code{numeric} representing
the segregation in the environment,
}
\details{
All functions have names end in "_ind". They take in input a 2xN matrix.
In the matix, each cell shows how many individuals from a given group are in
a certain unit. The rows of the matrix are considered the groups in the
environment. The columns are considered the different units. The functions
returns the value of the desired index for the environment as a
\code{numeric}.

At this point, all indices suppose that there are only two groups in the
environment. This will change later.
}
\section{Functions}{
\itemize{
\item \code{d_ind}: D Index

\item \code{gini_ind}: Gini index. It shoud be noticed that the Gini index requires more complex
calculations with respect to the other indices in the SISeg package. For this
reason, it takes more time to compute and the difference may be noticeable
in the construction of confidence intervals.

\item \code{mutual_info_ind}: Mutual Information Index. It uses the exponential as a
base for logarithms.

\item \code{theil_ind}: Theil Index (One of the possible normalization of the
Mutual Information Index). It uses the exponential as a
base for logarithms.

\item \code{atkinson_ind}: Atkinson Index as defined by Frankel and Volij (2011).
The weights are assumed to be uniform over the units.

\item \code{v_ind}: V Index, also known as Eta-squared index.

\item \code{isolation_ind}: Isolation Index. Notice, this is asymmetric.

\item \code{isolationInv_ind}: A convenience function to calculate the Isolation Index for
the second group in an environment.
}}

\examples{
env <- matrix(c(1,2,3,4,5,6,7,8), nrow = 2, byrow = TRUE)
d_ind(env)
gini_ind(env)
mutual_info_ind(env)
theil_ind(env)
atkinson_ind(env)
v_ind(env)
isolation_ind(env)
isolationInv_ind(env)
isolation_ind(env[c(2,1),]) == isolationInv_ind(env)
}
\references{
For the index formulas and properties, see discussion in (sources
are in order of relevance)

\describe{
\item{Atkinson Index}{
James, David R., and Karl E. Taeuber. 1985. ``Measures of Segregation.''
Sociological Methodology 15: 1–32.

Frankel, David M., and Oscar Volij. 2011. “Measuring School Segregation.”
Journal of Economic Theory 146 (1): 1–38.
}
\item{D Index}{Duncan, Otis Dudley, and Beverly Duncan. 1955. “A Methodological Analysis of
Segregation Indexes.” American Sociological Review 20 (2): 210–17.}
\item{Gini Index}{Hutchens, Robert. 1991. “Segregation Curves, Lorenz
Curves, and Inequality in the Distribution of People across Occupations.”
Mathematical Social Sciences 21 (1): 31 – 51.}
\item{Mutual Information Index and Theil Index}{
Mora, Ricardo, and Javier Ruiz-Castillo. 2011. “Entropy-Based Segregation
Indices.” Sociological Methodology 41 (1): 159–194.

Reardon, Sean F., and Glenn Firebaugh. 2002. “Measures of Multigroup
Segregation.” Sociological Methodology 32 (1): 33–67.

Frankel, David M., and Oscar Volij. 2011. “Measuring School Segregation.”
Journal of Economic Theory 146 (1): 1–38.
}
\item{V and Isolation Index}{
Bell, Wendell. 1954. “A Probability Model for the Measurement of Ecological
Segregation.” Social Forces 32 (4): 357–364.

Massey, Douglas S., and Nancy A. Denton. 1988. ``The Dimensions of
Residential Segregation.'' Social Forces 67 (2): 281–315.
}
}
}
\seealso{
Other indices: 
\code{\link{seg_ind_bin}}
}
\concept{indices}
