% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{seg_ind_bin}
\alias{seg_ind_bin}
\alias{d_ind_bin}
\alias{gini_ind_bin}
\alias{mutual_info_ind_bin}
\alias{theil_ind_bin}
\alias{atkinson_ind_bin}
\alias{v_ind_bin}
\alias{isolation_ind_bin}
\alias{isolationinv_ind_bin}
\title{Binomial Form of Segregation Indices}
\usage{
d_ind_bin(env, in_form = FALSE)

gini_ind_bin(env, in_form = FALSE)

mutual_info_ind_bin(env, in_form = FALSE)

theil_ind_bin(env, in_form = FALSE)

atkinson_ind_bin(env, in_form = FALSE)

v_ind_bin(env, in_form = FALSE)

isolation_ind_bin(env, in_form = FALSE)

isolationinv_ind_bin(env, in_form = FALSE)
}
\arguments{
\item{env}{2xN \code{matrix} of \code{numeric}.}

\item{in_form}{\code{logical}, is \code{env} in binomial form?}
}
\value{
\code{numeric} representing the index value for the data. This result
is identical to the basic segregation indices functions.
}
\description{
All these functions calculate the value of a segregation index in an
environment using the binomial form. They output a positive \code{numeric}
representing the segregation in the environment. These functions use the
binomial form of the indices. Results are identical to the use of standard
forms implemented in  \link[=seg_ind]{Segregation Indices}.
}
\details{
In general, it is possible to write segregation indices in two different
forms. The binomial form uses the proportion of a group in each unit, as
opposed to the standard form, which uses the proportion of each unit in the
group.It is usually faster and convenient to use the standard form, which is
indeed implemented in the basic functions to calculate segregation indices.

For a given index, the function calculating the index value through the
binomial form has a name that simply adds "_bin" to the basic function for
the index. For example, \code{d_ind()} and \code{d_ind_bin()}.
As for the base versions of the indices, these function take a 2xN matrix
as input. The rows of the matrix are considered the groups in the environment.
The columns are considered the different units. They returns the value of the
desired index for the environment as a \code{numeric}.

The input matrix can be in the so-called ``binomial form''. In this form,
the first row represents the proportion of individuals from group one in
a unit. The second row represents the proportions of individual from a
unit in the environment. If the input matrix is in this form, then the
argument \code{in_form} should be set to \code{TRUE}
}
\section{Functions}{
\itemize{
\item \code{d_ind_bin}: D Index

\item \code{gini_ind_bin}: Gini Index. As for its basic version, the Gini Index
is quite cumbersome to calculate.

\item \code{mutual_info_ind_bin}: Mutual Information Index.

\item \code{theil_ind_bin}: Theil index. (One of the possible normalization of the
Mutual Information Index)

\item \code{atkinson_ind_bin}: Atkinson Index as defined by Frankel and Volij (2011).
The weights are assumed to be uniform over the units.

\item \code{v_ind_bin}: V Index, also known as Eta-squared index.

\item \code{isolation_ind_bin}: Isolation Index. Notice, this is asymmetric.

\item \code{isolationinv_ind_bin}: A convenience function to calculate the Isolation Index for
the second group in an environment.
}}

\section{Naming convention}{

All functions have names ending in "_bin" and starting with the name of the
index. In the middle, they have the word "_ind". For example, "d_ind_bin".
Notice, everything is lowercase.
}

\examples{
env <- matrix(c(1,2,3,4,5,6,7,8), nrow = 2, byrow = TRUE)
basic <- d_ind(env)
bin <-  d_ind_bin(env)
all.equal(basic, bin)
env_bin_form <- env
env_bin_form[2,] <- colSums(env_bin_form)
env_bin_form[1,] <- env_bin_form[1,]/env_bin_form[2,]
env_bin_form[2,] <- env_bin_form[2,]/sum(env_bin_form[2,])
bin_bin_form <- d_ind_bin(env_bin_form, in_form = TRUE)
all.equal(bin_bin_form, bin)
}
\references{
The use of a binomial form is convenient, theoretically, to formulate
multi-group indices. See

Reardon, Sean F., and Glenn Firebaugh. 2002. “Measures of Multigroup
Segregation.” Sociological Methodology 32 (1): 33–67.
}
\seealso{
Other indices: 
\code{\link{seg_ind}}
}
\concept{indices}
