% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sigmas}
\alias{sigmas}
\alias{sigma_ig}
\alias{sigma_fm}
\alias{sigma_iu}
\title{Create the Var-Covar Matrix of a Framework from Data}
\usage{
sigma_ig(env)

sigma_fm(env)

sigma_iu(env)
}
\arguments{
\item{env}{2xk \code{matrix} of numeric}
}
\value{
A kxk Variance-Covariance \code{matrix}
}
\description{
These function produces the variance-covariance matrix for the statistical
frameworks starting from the data. The independence structures of the
frameworks are very different, resulting in more or less sparse
variance-covariance matrices.
}
\section{Functions}{
\itemize{
\item \code{sigma_ig}: Variance-covariance matrix for the independent-groups
framework

\item \code{sigma_fm}: Variance-covariance matrix for the full multinomial
framework

\item \code{sigma_iu}: Variance-covariance matrix for the independent-units
framework
}}

\examples{
env <- matrix(seq(8), nrow = 2, byrow = TRUE)
est1 <- sigma_ig(env)
est2 <- sigma_fm(env)
est3 <- sigma_iu(env)
}
