% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_index.R
\name{test_index}
\alias{test_index}
\title{Test Inference on New Indices}
\usage{
test_index(
  seg_index,
  n_b = 1e+05,
  asymp = TRUE,
  estimate = TRUE,
  difference = FALSE,
  framework = "all",
  graph = TRUE
)
}
\arguments{
\item{seg_index}{A \code{string} with the name of the index to be tested.}

\item{n_b}{\code{numeric} How many simulations in the tests? The number of
simulations in the estimate and difference tests will be capped to 1,000
if n_b is bigger than 1,000.}

\item{asymp}{\code{logical} Should the asymptotic tests be run?}

\item{estimate}{\code{logical} Should the estimate tests be run?}

\item{difference}{\code{logical} Should the difference tests be run?}

\item{framework}{\code{vector} or \code{all}. If vector, it should contain the
names of the framework to be tested.}

\item{graph}{\code{logical} Should the sampling and asymptotic distribution
be compared in a plot? Ignored if \code{asymp} is FALSE.}
}
\value{
\code{list} containing the results of the tests. If \code{asymp} is
\code{TRUE}, it will contain the results of the Kolgomorov-Smirnov test
against normality. A successful test will return a large p-value. If
\code{asymp} is \code{TRUE} and \code{graph} is also \code{TRUE},
the function will also produce plots
showing the sampling distribution for the tested index and the
asymptotic normal distribution for the index -- the two distribution
should be very similar. If \code{estimate} is \code{TRUE}, the
\code{list} will contain the coverage of the different methods for the
point-estimate. If \code{framework} is \code{TRUE}, the \code{list}
will contain the coverage of the different methods for the difference.
}
\description{
The user can define new segregation indices and SISeg will perform
inference on those automatically. After using the function
\code{\link{add_index}} to add a new index to the supported indices in
SISeg, \code{test_index} will perform automatic checks that the new
index and its related functions work as expected. The tests are based on
asymptotic theory and, for this reason, it will take time to run a full
battery of tests.
}
\details{
The inference techniques in SISeg are general and can be applied to any
(derivable) index. The user can specify a new index and then use the
inference functions in SISeg to get inference for the newly specified index.
The function \code{\link{add_index}} lets the user define new indices and
add them to SISeg list of supported indices. \code{test_index} tests that
the inference on the added index works as it should. It provides three
batteries of tests:
\describe{
\item{`asymp'}{Tests about the asymptotic normality of the index. Use these
to test that the derivative functions for the new index are correct.}
\item{estimate}{Tests to check that the confidence interval for the index
in a dataset works as expected. These tests check the coverage of the
bootstrap and asymptotic confidence intervals.}
\item{difference}{Tests the coverage of the confidence intervals for the
difference between indices in two environments. By default the
function will not run these tests. The argument is that if the
estimate tests succeed, the difference tests will also succeed.}
}

All tests are based on the inference performances on a very convenient case.
That is, the tests are perfomed by checking the behavior of the inferential
methods on a simple environment with 350,000 individuals and only four
columns. In this environment, asymptotic normality is reached for the
indices that SISeg natively supports and it should be reached for any
well-behaved function. Therefore, the tests consist in checking that
asymptotic normality holds for this environment and that the different
techniques reach the nominal coverage.

The user should be ware that running the entire set of tests will take few
hours, with the asymptotic tests only taking few minutes.
}
\examples{
# We can use the function to test indices that are already in SISeg
# Test the Theil index
# Asymptotic test are quick
test_index("Theil", estimate = FALSE, asymp = TRUE)
# Bootstrap tests are not. We only do 10 simulations
test_index("Theil", n_b = 10, asymp = FALSE, estimate = TRUE)
# The actual tests take a long time.
\donttest{test_index("Theil")}

}
